Beta 1.8.5.2545.0 -Added `UI Mode` slider to User Interface section, which allows to change between `Auto` (based on resolution, like before), `Full HD`, `HD` and `SD` (WIP PSVita Edition / low resolution mode) UI modes -Added Type Crafting Tabs manager, this will allow to customize the type crafting tabs from the Crafting Interface like other listings -Added (and modified) customization with UI Definitions to a lot of interfaces, although it's still incomplete and subject to change -Added `sizeable` tab icon type, meant to be used for icon changes between UI modes -Added `put_tab_state_offset` UI Definition Element type -Enhanced interface navigation, now slots will behave like widgets, also allowing cursor navigation on widgets, even using keyboard, depending on the screen -Now, resource managers based on IDs, such as Crafting Tab Listing, Creative Tab Listing, Loom Tab Listing, Biome Overrides and Control Types will be more consistent and allow modifications without problems -Now, the PS Vita control icons will have a more accurate size -Now, the `HUD Distance` slider will be applied to the `Animated Character` -Now, the save storage bar in the Play Game screen will be displayed only in the FHD mode -Now, Control Types will allow fonts for different UI modes -Now, Loom Tab Listing will allow to change the first tab icon -Now, Creative Tab Listing will allow to change the search tab icon -Now, Legacy Sliders will apply the value changes only when released after dragging using a mouse -Fixed sun and moon textures not being loaded after toggling (and loading) the `Legacy Sky` option -Fixed boat left stick deadzone being applied to the cursor -Fixed Custom Crafting not showing the selected slot correctly -Fixed Mojangles 11 not displaying LE icons correctly -Fixed being able to navigate on the screen before the intro is finished or skipped -Fixed controller cursor limit not allowing to use the recipe book -Reformatted code to be more consistent

+++ src/main/java/wily/legacy/mixin/base/client/LegacyLoadingMessageScreenMixin.java
@@ -37,12 +37,12 @@ public class LegacyLoadingMessageScreenMixin extends Screen implements LegacyLoa
     //?}
 
     @Override
-    public int getProgress() {
+    public float getProgress() {
         return legacyLoadingScreen.getProgress();
     }
 
     @Override
-    public void setProgress(int progress) {
+    public void setProgress(float progress) {
         legacyLoadingScreen.setProgress(progress);
     }
